requirements.txt
fastapi
Pillow
python-multipart
torch
transformers
uvicorn
moviepy